Epiphone regarding branches
Made a minor edit to the main README.md check list.
I may have imported a revised template for the master document.
Apparently made a minor change to the quick start txt file.
Imported a revised template for the quick start and am in the middle of formatting the contents to reflect the new template.

In the middle of editing the quick start file I realized that I should probably be branching each document file to edit them individually.